Ensure we retain the original InetSocketAddress when connect to a remote peer when using epoll transport.
Motivation:

We should retain the original InetSocketAddress when connecto to a remote peer so the user can still query the origin hostname if getHostString() is used.

Modifications:

Store a reference to the origin InetSocketAddress and use it once the chnannel becomes active.

Result:

Same behavior when using epoll transport and nio transport.